Fix for the broken fix of the compile warnings that I fixed yesterday on 64bit platforms, but ended up needing to refix on 32bit platforms.
The universe truly hates me :P


git-svn-id: http://svn.ratbox.org/svnroot/trunk@7916 b93f080a-4cfa-0310-a19a-80a68f531ef9

+++ include/ircd.h
@@ -102,7 +102,7 @@ extern dlink_list lazylink_nicks;
 extern int callbacks_called;
 
 extern int make_daemon(void);
-extern size_t get_maxrss(void);
+extern unsigned long get_maxrss(void);
 extern void set_time(void);
 
 #endif
